requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
mediapipe==0.10.11
pandas
numpy
plotly
scipy
reportlab
python-docx
python-pptx
duckduckgo-search
SpeechRecognition
PyAudio
pydub
